Moved clock network base class and its children to a sperate file
